fix: rename RequestType enum members
Move functions from otlp_http  to mod
