--- src/service/traces/otlp_http.rs
@@ -1,77 +1,0 @@
-// Copyright 2024 Zinc Labs Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-use std::io::Error;
-
-use actix_web::{http, web, HttpResponse};
-use opentelemetry_proto::tonic::collector::trace::v1::ExportTraceServiceRequest;
-use prost::Message;
-
-use crate::common::meta::http::HttpResponse as MetaHttpResponse;
-
-pub async fn traces_proto(
-    org_id: &str,
-    body: web::Bytes,
-    in_stream_name: Option<&str>,
-) -> Result<HttpResponse, Error> {
-    let request = match ExportTraceServiceRequest::decode(body) {
-        Ok(v) => v,
-        Err(e) => {
-            return Ok(HttpResponse::BadRequest().json(MetaHttpResponse::error(
-                http::StatusCode::BAD_REQUEST.into(),
-                format!("Invalid proto: {}", e),
-            )));
-        }
-    };
-    super::handle_trace_request(
-        org_id,
-        request,
-        super::RequestType::Protobuf,
-        in_stream_name,
-    )
-    .await
-}
-
-pub async fn traces_json(
-    org_id: &str,
-    body: web::Bytes,
-    in_stream_name: Option<&str>,
-) -> Result<HttpResponse, Error> {
-    let request = match serde_json::from_slice::<ExportTraceServiceRequest>(body.as_ref()) {
-        Ok(req) => req,
-        Err(e) => {
-            return Ok(HttpResponse::BadRequest().json(MetaHttpResponse::error(
-                http::StatusCode::BAD_REQUEST.into(),
-                format!("Invalid json: {}", e),
-            )));
-        }
-    };
-    super::handle_trace_request(org_id, request, super::RequestType::Http, in_stream_name).await
-}
-
-#[cfg(test)]
-mod tests {
-    use config::utils::json::json;
-
-    use crate::service::ingestion::grpc::get_val_for_attr;
-
-    #[test]
-    fn test_get_val_for_attr() {
-        let in_val = 10.00;
-        let input = json!({ "key": in_val });
-        let resp = get_val_for_attr(input);
-        assert_eq!(resp.as_str().unwrap(), in_val.to_string());
-    }
-}
